Tickets and Contracts now belong to Company (+ long -> int fixes)
If a customer (=contactperson) is removed but the company still exists,
then the tickets would be gone. Now the tickets and contracts belong to
a company.

//todo
- clean up of code after this change
- fix test classes

//question
why did we change from long to int for our PKs? was it that important?

+++ src/domain/facades/UserFacade.java
@@ -86,7 +86,7 @@ public class UserFacade extends Facade {
 	 */
 
 	// We hadn't time to implement this feature in the GUI
-	public void registerCustomerUsingExistingCompany(String username, String password, String firstName, String lastName, Long companyId) throws InformationRequiredException {
+	public void registerCustomerUsingExistingCompany(String username, String password, String firstName, String lastName, int companyId) throws InformationRequiredException {
 		// check to see if signed in user is Admin
 		actemium.checkPermission(EmployeeRole.ADMINISTRATOR);
 		actemium.existingUsername(username);
